+++ core-notifications-fe/src/core-notifications/mock.ts
@@ -3,7 +3,7 @@ import { faker } from '@faker-js/faker'
 const simpleNotification = (userFn: Function) => {
   const user = userFn()
 
-  return ({
+  return {
     kind: 'Simple',
 
     to_id: user.id,
@@ -12,8 +12,8 @@ const simpleNotification = (userFn: Function) => {
     read: faker.datatype.boolean(),
 
     title: faker.lorem.sentence(),
-    description: faker.lorem.sentences(2)
-  })
+    description: faker.lorem.sentences(2),
+  }
 }
 
 const taskNotification = (userFn: Function) => {
@@ -27,47 +27,56 @@ const taskNotification = (userFn: Function) => {
 
     title: faker.lorem.sentence(),
     description: faker.lorem.sentences(2),
-    actions: [{label: 'Accept'}, {label: 'Decline'}]
+    actions: [
+      { label: 'Accept', url: '#accept' },
+      { label: 'Decline', url: '#decline' },
+    ],
   }
 }
 
-// const groupNotification = () => ({
-//   kind: 'Group',
-//   title: faker.lorem.sentence(),
-//   description: faker.lorem.sentences(2)
-// })
-
-
 const globalNotification = (_userFn: Function) => {
   const publish_at = faker.helpers.arrayElement([null, faker.date.anytime()])
-  const expire_at = publish_at ? faker.date.future({refDate: publish_at}) : faker.date.anytime()
+  const expire_at = publish_at
+    ? faker.date.future({ refDate: publish_at })
+    : faker.date.anytime()
 
   return {
-  kind: 'Global',
-  read_bloom: [],
-  publish_at: publish_at,
-  expire_at: expire_at,
-  read: faker.number.int({max: 2000000}),
-  title: faker.lorem.sentence(),
-  description: faker.lorem.sentences(2),
-  link: faker.internet.url()
-}
+    kind: 'Global',
+    scope: faker.helpers.arrayElement([
+      null,
+      'active',
+      ".joins(:leaderships).where(leaderships: {status: 'active', org_unit_id: OrgUnit.where(state: 'GA')})",
+    ]),
+    read_bloom: [],
+    publish_at: publish_at,
+    expire_at: expire_at,
+    read: faker.number.int({ max: 2000000 }),
+    title: faker.lorem.sentence(),
+    description: faker.lorem.sentences(2),
+    link: faker.helpers.arrayElement([
+      null,
+      faker.internet.url(),
+      { url: faker.internet.url(), label: faker.lorem.words(2) },
+    ]),
+  }
 }
 
 const mockNotification = (userFn: Function) => {
-  const kind = faker.helpers.arrayElement([globalNotification, simpleNotification, taskNotification])
+  const kind = faker.helpers.arrayElement([
+    globalNotification,
+    simpleNotification,
+    taskNotification,
+  ])
 
-  return (
-    {
-      ...kind(userFn),
+  return {
+    ...kind(userFn),
 
-      id: faker.string.uuid(),
-      created_at: faker.date.recent(),
-      updated_at: faker.date.recent(),
-    })
+    id: faker.string.uuid(),
+    created_at: faker.date.recent(),
+    updated_at: faker.date.recent(),
+  }
 }
 
-export const mockNotifications = (count = 35, userFn: Function)  => {
-  return Array.from({length: count}, () => mockNotification(userFn))
+export const mockNotifications = (count = 35, userFn: Function) => {
+  return Array.from({ length: count }, () => mockNotification(userFn))
 }
-