+++ core-notifications-fe/src/core-notifications/resources/notifications/TableList.tsx
@@ -1,6 +1,7 @@
 import {
   PolymorphicReferenceField,
   TimeAgoField,
+  ToggleButtonInput,
 } from '@moonlight-labs/core-base-fe'
 import {
   ChipField,
@@ -26,7 +27,8 @@ import {
 const adminFilters = [
   // <Typography variant="h4" alwaysOn>Web Hooks<br/></Typography>,
   <TextInput label="Search" source="q" alwaysOn />,
-  // <ToggleButtonInput label="Read" source="read" />,
+  <ToggleButtonInput alwaysOn source="kind" choices={[{id: 'Global', name: 'Global'}, {id: 'Task', name: 'Task'}, {id: 'Simple', name: 'Simple'}]} />,
+  <ToggleButtonInput alwaysOn source="status" choices={[{id: 'draft', name: 'Draft'}, {id: 'upcoming', name: 'Upcoming'}, {id: 'Live', name: 'Live'}, {id: 'expired', name: 'Expired'}]} />,
 ]
 
 const userFilters = [