+++ core-webhooks-fe/src/core-webhooks/resources/CodeField.tsx
@@ -1,4 +1,5 @@
 import { Box } from '@mui/material'
+import { get } from 'lodash'
 import { FieldProps, useRecordContext } from 'react-admin'
 // import { LightAsync as SyntaxHighlighter } from 'react-syntax-highlighter'
 
@@ -14,7 +15,10 @@ export const CodeField = ({ label, source }: FieldProps) => {
   const record = useRecordContext()
   if (!record || !source) return null
 
-  const code = JSON.stringify(record[source], null, 2) || ''
+  const value = get(record, source, null)
+  if(!value) return null
+
+  const code = JSON.stringify(value, null, 2) || ''
 
   return (
     <Box sx={{ border: 1, borderColor: 'grey.500', borderRadius: '6px' }}>