+++ core-webhooks-fe/src/core-webhooks/resources/webhooks/index.ts
@@ -1,10 +1,12 @@
 import WebhookIcon from '@mui/icons-material/Webhook'
 import { WebhookTable } from './TableList'
 import { WebhookShow } from './show'
+import { RaRecord } from 'react-admin'
 
 export class Webhooks {
   static Name = 'Webhook'
   static Icon = WebhookIcon
   static List = WebhookTable
   static Show = WebhookShow
+  static recordRepresentation = (webhook: RaRecord ) => `${webhook.source} / ${webhook.event}`
 }