+++ core-notifications-fe/CHANGELOG.md
@@ -1,1 +1,9 @@
 # @moonlight-labs/core-notifications-fe
+
+## 0.2.8
+
+### Patch Changes
+
+- Base enhancements, Webhooks improvements
+- Updated dependencies
+  - @moonlight-labs/core-base-fe@0.2.7