+++ core-versions-fe/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @moonlight-labs/core-versions-fe
 
+## 0.2.7
+
+### Patch Changes
+
+- Base enhancements, Webhooks improvements
+- Updated dependencies
+  - @moonlight-labs/core-base-fe@0.2.7
+
 ## 0.2.6
 
 ### Patch Changes